Fixing nullability

+++ Tests/Unit/NetworkServer/BasicsStation/ClientCertificateValidatorServiceTests.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+#nullable enable
+
 namespace LoRaWan.Tests.Unit.NetworkServer.BasicsStation
 {
     using System;
@@ -20,38 +22,40 @@ namespace LoRaWan.Tests.Unit.NetworkServer.BasicsStation
         private readonly Mock<IBasicsStationConfigurationService> stationConfigService;
         private readonly Mock<ILogger<ClientCertificateValidatorService>> logger;
         private const string InvalidStationEui = "NotAStationEui";
-        private string clientCertificateWithEuiPath;
-        private string clientCertificateWithInvalidEuiPath;
+        private string? clientCertificateWithEuiPath;
 
         public ClientCertificateValidatorServiceTests()
         {
-            stationConfigService = new Mock<IBasicsStationConfigurationService>();
-            logger = new Mock<ILogger<ClientCertificateValidatorService>>();
-            clientCertValidatorSvc = new ClientCertificateValidatorService(stationConfigService.Object, logger.Object);
-        }
-
-        [Fact]
-        public async Task ValidateAsync_Throws_WithoutCertificate()
-        {
-            using var chain = X509Chain.Create();
-            await Assert.ThrowsAsync<ArgumentNullException>(() => this.clientCertValidatorSvc.ValidateAsync(null, chain, System.Net.Security.SslPolicyErrors.None, default));
+            this.stationConfigService = new Mock<IBasicsStationConfigurationService>();
+            this.logger = new Mock<ILogger<ClientCertificateValidatorService>>();
+            this.clientCertValidatorSvc = new ClientCertificateValidatorService(this.stationConfigService.Object, this.logger.Object);
         }
 
         [Fact]
         public async Task ValidateAsync_Throws_WithoutChain()
         {
+            if (string.IsNullOrEmpty(this.clientCertificateWithEuiPath))
+            {
+                throw new InvalidOperationException("Client certificate was not properly initialized.");
+            }
+
             using var cert = new X509Certificate2(this.clientCertificateWithEuiPath);
-            await Assert.ThrowsAsync<ArgumentNullException>(() => this.clientCertValidatorSvc.ValidateAsync(cert, null, System.Net.Security.SslPolicyErrors.None, default));
+            _ = await Assert.ThrowsAsync<ArgumentNullException>(() => this.clientCertValidatorSvc.ValidateAsync(cert, null, System.Net.Security.SslPolicyErrors.None, default));
         }
 
         [Fact]
         public async Task ValidateAsync_ReturnsTrue_WithExpectedThumbprint()
         {
+            if (string.IsNullOrEmpty(this.clientCertificateWithEuiPath))
+            {
+                throw new InvalidOperationException("Client certificate was not properly initialized.");
+            }
+
             using var cert = new X509Certificate2(this.clientCertificateWithEuiPath);
             using var chain = X509Chain.Create();
 
-            stationConfigService.Setup(x => x.GetAllowedClientThumbprintsAsync(It.IsAny<StationEui>(), It.IsAny<CancellationToken>()))
-                .Returns(Task.FromResult(new[] { cert.Thumbprint }));
+            _ = this.stationConfigService.Setup(x => x.GetAllowedClientThumbprintsAsync(It.IsAny<StationEui>(), It.IsAny<CancellationToken>()))
+                                         .Returns(Task.FromResult(new[] { cert.Thumbprint }));
 
             var result = await this.clientCertValidatorSvc.ValidateAsync(cert, chain, System.Net.Security.SslPolicyErrors.None, default);
 
@@ -61,11 +65,16 @@ namespace LoRaWan.Tests.Unit.NetworkServer.BasicsStation
         [Fact]
         public async Task ValidateAsync_ReturnsFalse_WithNonExpectedThumbprint()
         {
+            if (string.IsNullOrEmpty(this.clientCertificateWithEuiPath))
+            {
+                throw new InvalidOperationException("Client certificate was not properly initialized.");
+            }
+
             using var cert = new X509Certificate2(this.clientCertificateWithEuiPath);
             using var chain = X509Chain.Create();
 
-            stationConfigService.Setup(x => x.GetAllowedClientThumbprintsAsync(It.IsAny<StationEui>(), It.IsAny<CancellationToken>()))
-                .Returns(Task.FromResult(new[] { "AnotherThumbprint" }));
+            _ = this.stationConfigService.Setup(x => x.GetAllowedClientThumbprintsAsync(It.IsAny<StationEui>(), It.IsAny<CancellationToken>()))
+                                         .Returns(Task.FromResult(new[] { "AnotherThumbprint" }));
 
             var result = await this.clientCertValidatorSvc.ValidateAsync(cert, chain, System.Net.Security.SslPolicyErrors.None, default);
 
@@ -75,47 +84,61 @@ namespace LoRaWan.Tests.Unit.NetworkServer.BasicsStation
         [Fact]
         public async Task ValidateAsync_ReturnsFalse_WithInvalidCNCertificate()
         {
-            using var cert = new X509Certificate2(this.clientCertificateWithInvalidEuiPath);
-            using var chain = X509Chain.Create();
+            var clientCertificateWithInvalidEuiPath = await BasicsStationNetworkServerTests.CreatePfxCertificateAsync(false, InvalidStationEui);
+            try
+            {
+                using var cert = new X509Certificate2(clientCertificateWithInvalidEuiPath);
+                using var chain = X509Chain.Create();
 
-            stationConfigService.Setup(x => x.GetAllowedClientThumbprintsAsync(It.IsAny<StationEui>(), It.IsAny<CancellationToken>()))
-                .Returns(Task.FromResult(new[] { cert.Thumbprint }));
+                _ = this.stationConfigService.Setup(x => x.GetAllowedClientThumbprintsAsync(It.IsAny<StationEui>(), It.IsAny<CancellationToken>()))
+                                             .Returns(Task.FromResult(new[] { cert.Thumbprint }));
 
-            var result = await this.clientCertValidatorSvc.ValidateAsync(cert, chain, System.Net.Security.SslPolicyErrors.None, default);
+                var result = await this.clientCertValidatorSvc.ValidateAsync(cert, chain, System.Net.Security.SslPolicyErrors.None, default);
 
-            Assert.False(result);
-            Assert.Contains(this.logger.Invocations, i => i.Arguments.Any(a => a.ToString().Contains(InvalidStationEui, System.StringComparison.OrdinalIgnoreCase)));
+                Assert.False(result);
+                Assert.Contains(this.logger.Invocations, i => i.Arguments.Any(a => a.ToString()!.Contains(InvalidStationEui, StringComparison.OrdinalIgnoreCase)));
+            }
+            finally
+            {
+                if (!string.IsNullOrEmpty(clientCertificateWithInvalidEuiPath))
+                    File.Delete(clientCertificateWithInvalidEuiPath);
+            }
+
         }
 
         [Fact]
         public async Task ValidateAsync_ReturnsTrue_WithUntrustedRoot()
         {
+            if (string.IsNullOrEmpty(this.clientCertificateWithEuiPath))
+            {
+                throw new InvalidOperationException("Client certificate was not properly initialized.");
+            }
+
             using var cert = new X509Certificate2(this.clientCertificateWithEuiPath);
             using var chain = X509Chain.Create();
-            chain.Build(cert);
+            _ = chain.Build(cert);
 
-            stationConfigService.Setup(x => x.GetAllowedClientThumbprintsAsync(It.IsAny<StationEui>(), It.IsAny<CancellationToken>()))
-                .Returns(Task.FromResult(new[] { cert.Thumbprint }));
+            _ = this.stationConfigService.Setup(x => x.GetAllowedClientThumbprintsAsync(It.IsAny<StationEui>(), It.IsAny<CancellationToken>()))
+                                         .Returns(Task.FromResult(new[] { cert.Thumbprint }));
 
             var result = await this.clientCertValidatorSvc.ValidateAsync(cert, chain, System.Net.Security.SslPolicyErrors.None, default);
 
             Assert.True(result);
-            Assert.Contains(this.logger.Invocations, i => i.Arguments.Any(a => a.ToString().Contains(X509ChainStatusFlags.UntrustedRoot.ToString(), System.StringComparison.OrdinalIgnoreCase)));
+            Assert.Contains(this.logger.Invocations, i => i.Arguments.Any(a => a.ToString()!.Contains(X509ChainStatusFlags.UntrustedRoot.ToString(), StringComparison.OrdinalIgnoreCase)));
         }
 
 
 
         public Task DisposeAsync()
         {
-            File.Delete(this.clientCertificateWithEuiPath);
-            File.Delete(this.clientCertificateWithInvalidEuiPath);
+            if (!string.IsNullOrEmpty(this.clientCertificateWithEuiPath))
+                File.Delete(this.clientCertificateWithEuiPath);
             return Task.CompletedTask;
         }
 
         public async Task InitializeAsync()
         {
-            this.clientCertificateWithEuiPath = await BasicsStationNetworkServerTests.CreatePfxCertificate(false);
-            this.clientCertificateWithInvalidEuiPath = await BasicsStationNetworkServerTests.CreatePfxCertificate(false, InvalidStationEui);
+            this.clientCertificateWithEuiPath = await BasicsStationNetworkServerTests.CreatePfxCertificateAsync(false);
         }
     }
 }
